requirements.txt
streamlit
opencv-python
pytesseract
numpy
pandas
pillow
